+Константы карты и ее инициализация зане с в отдельный модуль init-map
